Move all source files to src directory.
