+++ gulpfile.js
@@ -21,7 +21,7 @@ gulp.task('lint', function () {
 });
 
 gulp.task('build-client', ['lint', 'move-client'], function () {
-  return gulp.src(['client/js/app.js'])
+  return gulp.src(['src/client/js/app.js'])
     .pipe(webpack(require('./webpack.config.js')))
     //.pipe(uglify())
     .pipe(gulp.dest('bin/client/js/'));
@@ -34,22 +34,22 @@ gulp.task('move-client', function () {
 
 
 gulp.task('build-server', ['lint'], function () {
-  return gulp.src(['server/**/*.*', 'server/**/*.js'])
+  return gulp.src(['src/server/**/*.*', 'src/server/**/*.js'])
     .pipe(gulp.dest('bin/server/'));
 });
 
 gulp.task('watch', ["build"], function () {
-  gulp.watch('client/**/*.*', ['build-client', 'move-client']);
-  gulp.watch('server/*.*', 'server/**/*.js', ['build-server']);
+  gulp.watch('src/client/**/*.*', ['build-client', 'move-client']);
+  gulp.watch('src/server/*.*', 'src/server/**/*.js', ['build-server']);
   gulp.start("run");
 });
 
 gulp.task('run', ["build"], function () {
     nodemon({
         delay: 10,
-        script: 'server/server.js',
+        script: './server/server.js',
         cwd: "./bin/",
-        args: ["/server/config.json"],
+        args: ["config.json"],
         ext: 'html js css'
     })
     .on('restart', function () {
